LayoutTests/fast/events/wheel/wheelevent-mousewheel-interaction-expected.txt

Tests the interaction between the standard 'wheel' event and the non-standard 'mousewheel' one

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Standard wheel event was fired.
PASS testEvent.__proto__ is WheelEvent.prototype
PASS successfullyParsed is true

TEST COMPLETE
TOP TOP TOP TOP TOP TOP TOP Scroll mouse wheel over here Scroll mouse wheel over here Scroll mouse wheel over here Scroll mouse wheel over here Scroll mouse wheel over here Scroll mouse wheel over here END END END END END END END
